NewsTailorBE/requirements.txt
asgiref==3.8.1
Django==4.2.16
django-environ==0.11.2
djangorestframework==3.15.2
django-cors-headers
djangorestframework-simplejwt~=5.3.1
PyJWT
pytz
psycopg2-binary
sqlparse==0.5.1
typing_extensions==4.12.2
requests~=2.32.3
python-dotenv~=1.0.1
feedparser~=6.0.11
Flask~=3.0.3
pytest~=8.3.3
selenium~=4.26.1
openai~=1.54.5
beautifulsoup4~=4.12.3

webdriver-manager~=4.0.2